broken
fk me dude

+++ bengine/headers/model.h
@@ -71,7 +71,7 @@ public:
 			if (name == "texture_diffuse")  number = std::to_string(diffuseNr++);
 			else if (name == "texture_specular") number = std::to_string(specularNr++);
 
-			shader.setFloat(("material." + name + number).c_str(), i);
+			shader.setInt(("material." + name + number).c_str(), i);
 			glBindTexture(GL_TEXTURE_2D, textures[i].id);
 
 		}
